requirements.txt
aiofiles==22.1.0
bcrypt==4.0.1
fastapi==0.88.0
flake8==6.0.0
passlib==1.7.4
pytest==7.2.0
python-jose==3.3.0
python-multipart==0.0.5
sqlalchemy==1.4.45
uvicorn==0.20.0